requirements.txt
appdirs==1.4.4
beautifulsoup4==4.14.3
blinker==1.9.0
boolean.py==5.0
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.2.0
click==8.3.1
colorama==0.4.6
curl_cffi==0.13.0
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
filelock==3.25.0
finta==1.3
Flask==3.1.3
frozendict==2.4.7
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jugaad-data==0.24
license-expression==30.4.4
markdown-it-py==4.0.0
markup==0.2
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
multitasking==0.0.12
numpy==2.4.2
packageurl-python==0.17.6
packaging==26.0
pandas==3.0.1
peewee==4.0.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.9.2
protobuf==6.33.5
py-serializable==2.1.0
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.32.5
rich==14.3.3
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tomli==2.4.0
tomli_w==1.2.0
typing_extensions==4.15.0
tzdata==2023.3
urllib3==2.6.3
waitress==3.0.2
websockets==16.0
Werkzeug==3.1.6
yfinance==1.2.0
google-genai
google-api-python-client
google-auth-oauthlib
python-telegram-bot
requests
